xdp_consumer/requirements.txt
pika
pyyaml 
pyjson 
pickledb 
netaddr 
pytime
